backend/requirements.txt
Flask
pymongo
flask_pymongo
Pyrebase4
flask_cors
python-dotenv
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
motor==3.3.1